requirements.txt
aiohttp
asyncio
asyncpg
nextcord
python-dotenv
pytz
quart
requests